Twonkie v2.0 alpha hardware
Upgrade the project to KiCad 6.0, fixing a whole bunch of new ERC/DRC
errors and warnings.

Switch from INA260 to INA237 which supports EPR voltage ranges, with
external shunt resistors.

Strap a previously unused pin on the STM32 to indicate the board
revision, so firmware can detect which hardware it's running on.